requirements.txt
python-chess
stockfish
pandas
matplotlib
scikit-learn